Initial commit
Will not print the list of choices that were made. 
Seems like it is tracking it, but then fails to print it out on the GUI.
Needs styled.
Needs to print the recipes. Will need to create a function that compares the full list then picks a recipe based off of that.